requirements.txt
torch==2.5.0
torchvision==0.20.0
pillow==10.0.0
streamlit==1.40.0
numpy==1.24.3
pandas==1.5.3
matplotlib==3.6.0